# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisResult": {
        "additionalProperties": false,
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ScanError"
            },
            "title": "Errors",
            "type": "array"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "scan": {
            "$ref": "#/components/schemas/ScanMetadata"
          },
          "schema_version": {
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "screenshots": {
            "items": {
              "$ref": "#/components/schemas/ScreenshotRef"
            },
            "title": "Screenshots",
            "type": "array"
          },
          "sections": {
            "$ref": "#/components/schemas/SectionSet"
          },
          "target": {
            "$ref": "#/components/schemas/TargetInfo"
          }
        },
        "required": [
          "scan",
          "target",
          "sections"
        ],
        "title": "AnalysisResult",
        "type": "object"
      },
...
      "ErrorCode": {
        "description": "Machine-readable error identity, shared by problem responses and scan errors.",
        "enum": [
          "INVALID_REQUEST",
          "INVALID_URL",
          "BLOCKED_TARGET",
          "ROBOTS_DISALLOWED",
          "DNS_FAILURE",
          "CONNECT_FAILURE",
          "TLS_FAILURE",
          "NAVIGATION_TIMEOUT",
          "BROWSER_UNAVAILABLE",
          "ACCESS_RESTRICTED",
          "BUDGET_EXHAUSTED",
          "MISSING_EVIDENCE",
          "ANALYZER_FAILED",
          "ANALYZER_TIMEOUT",
          "SCAN_NOT_FOUND",
          "RESULT_EXPIRED",
          "SCAN_IN_PROGRESS",
          "RATE_LIMITED",
          "AI_DISABLED",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": false,
        "description": "RFC 9457 problem document.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Stable URI identifying the problem kind.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "RunContext": {
        "additionalProperties": false,
        "description": "The conditions the measurements were taken under.\n\nPerformance and design numbers are meaningless without this, so it travels with every\nresult and is printed in every report.",
        "properties": {
          "browser_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Browser Name"
          },
          "browser_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Browser Version"
          },
          "collection_mode": {
            "description": "'http_only' or 'browser'. Tells the reader which evidence base was used.",
            "title": "Collection Mode",
            "type": "string"
          },
          "cpu_throttling": {
            "default": "none",
            "title": "Cpu Throttling",
            "type": "string"
          },
          "device_scale_factor": {
            "default": 1.0,
            "title": "Device Scale Factor",
            "type": "number"
          },
          "locale": {
            "default": "en-US",
            "title": "Locale",
            "type": "string"
          },
          "network_throttling": {
            "default": "none",
            "title": "Network Throttling",
            "type": "string"
          },
          "settle_reached": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settle Reached"
          },
          "timezone": {
            "default": "UTC",
            "title": "Timezone",
            "type": "string"
          },
          "user_agent": {
            "default": "WebLens/0.1.0 (+https://github.com/weblens; passive website analyzer; respects robots.txt)",
            "title": "User Agent",
            "type": "string"
          },
          "viewport": {
            "$ref": "#/components/schemas/Viewport"
          },
          "wait_strategy": {
            "description": "e.g. 'domcontentloaded+network_idle(5000ms)'.",
            "title": "Wait Strategy",
            "type": "string"
          }
        },
        "required": [
          "wait_strategy",
          "collection_mode"
        ],
        "title": "RunContext",
        "type": "object"
      },
...
      "ScanAcceptedResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Links",
            "type": "object"
          },
          "normalized_url": {
            "title": "Normalized Url",
            "type": "string"
          },
          "requested_url": {
            "title": "Requested Url",
            "type": "string"
          },
          "scan_id": {
            "title": "Scan Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ScanStatus"
          }
        },
        "required": [
          "scan_id",
          "status",
          "requested_url",
          "normalized_url",
          "created_at",
          "links"
        ],
        "title": "ScanAcceptedResponse",
        "type": "object"
      },
      "ScanError": {
        "additionalProperties": false,
        "description": "A failure recorded on a result rather than raised to the caller.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "scope": {
            "enum": [
              "scan",
              "stage",
              "analyzer"
            ],
            "title": "Scope",
            "type": "string"
          },
          "subject": {
            "description": "Stage key or analyzer id the failure belongs to.",
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "code",
          "scope",
          "subject",
          "message"
        ],
        "title": "ScanError",
        "type": "object"
      },
      "ScanJobState": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "problem": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProblemDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "$ref": "#/components/schemas/StageProgress"
          },
          "requested_url": {
            "title": "Requested Url",
            "type": "string"
          },
          "scan_id": {
            "title": "Scan Id",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageRun"
            },
            "title": "Stages",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/ScanStatus"
          }
        },
        "required": [
          "scan_id",
          "status",
          "requested_url"
        ],
        "title": "ScanJobState",
        "type": "object"
      },
...
      "ScanMetadata": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "engine_version": {
            "default": "0.1.0",
            "title": "Engine Version",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "options": {
            "$ref": "#/components/schemas/ScanOptions"
          },
          "run_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "scan_id": {
            "title": "Scan Id",
            "type": "string"
          },
          "schema_version": {
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageRun"
            },
            "title": "Stages",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/ScanStatus"
          }
        },
        "required": [
          "scan_id",
          "status",
          "created_at",
          "options"
        ],
        "title": "ScanMetadata",
        "type": "object"
      },
      "ScanOptions": {
        "additionalProperties": false,
        "description": "Per-scan options. ``extra='forbid'`` so an option typo is a 422, not a silent no-op.",
        "properties": {
          "include_full_page_screenshot": {
            "default": false,
            "title": "Include Full Page Screenshot",
            "type": "boolean"
          },
          "include_screenshot": {
            "default": true,
            "title": "Include Screenshot",
            "type": "boolean"
          },
          "responsive_widths": {
            "items": {
              "type": "integer"
            },
            "maxItems": 6,
            "title": "Responsive Widths",
            "type": "array"
          },
          "sections": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SectionKey"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Restrict analysis to these sections. None means all.",
            "title": "Sections"
          },
          "viewport": {
            "$ref": "#/components/schemas/Viewport"
          }
        },
        "title": "ScanOptions",
        "type": "object"
      },
      "ScanRequest": {
        "additionalProperties": false,
        "properties": {
          "options": {
            "$ref": "#/components/schemas/ScanOptions"
          },
          "url": {
            "description": "Public http(s) URL. Normalized and guarded server-side.",
            "maxLength": 2048,
            "minLength": 3,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "ScanRequest",
        "type": "object"
      },
      "ScanStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "completed_with_errors",
          "failed",
          "cancelled"
        ],
        "title": "ScanStatus",
        "type": "string"
      },
      "ScreenshotRef": {
        "additionalProperties": false,
        "description": "Screenshot metadata plus inline data.\n\nThe client moves the bytes into its own IndexedDB store as a Blob and does not keep them\nin the result record, which is why this is a separate top-level list rather than being\nburied inside the design payload.",
        "properties": {
          "data_base64": {
            "title": "Data Base64",
            "type": "string"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "mime_type": {
            "default": "image/png",
            "title": "Mime Type",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "label",
          "width",
          "height",
          "data_base64"
        ],
        "title": "ScreenshotRef",
        "type": "object"
      },
      "SectionKey": {
        "description": "The eight report sections. Order here is the canonical presentation order.",
        "enum": [
          "design",
          "technology",
          "security",
          "performance",
          "accessibility",
          "seo",
          "architecture",
          "network"
        ],
        "title": "SectionKey",
        "type": "string"
      },
...
      "SectionSet": {
        "additionalProperties": false,
        "description": "All eight sections, as explicit fields so both sides of the wire stay typed.",
        "properties": {
          "accessibility": {
            "$ref": "#/components/schemas/Section_AccessibilityPayload_"
          },
          "architecture": {
            "$ref": "#/components/schemas/Section_ArchitecturePayload_"
          },
          "design": {
            "$ref": "#/components/schemas/Section_DesignPayload_"
          },
          "network": {
            "$ref": "#/components/schemas/Section_NetworkPayload_"
          },
          "performance": {
            "$ref": "#/components/schemas/Section_PerformancePayload_"
          },
          "security": {
            "$ref": "#/components/schemas/Section_SecurityPayload_"
          },
          "seo": {
            "$ref": "#/components/schemas/Section_SeoPayload_"
          },
          "technology": {
            "$ref": "#/components/schemas/Section_TechnologyPayload_"
          }
        },
        "required": [
          "design",
          "technology",
          "security",
          "performance",
          "accessibility",
          "seo",
          "architecture",
          "network"
        ],
        "title": "SectionSet",
        "type": "object"
      },
...
      "Section_AccessibilityPayload_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccessibilityPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "interpretations": {
            "items": {
              "$ref": "#/components/schemas/Interpretation"
            },
            "title": "Interpretations",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/SectionMeta"
          }
        },
        "required": [
          "meta"
        ],
        "title": "Section[AccessibilityPayload]",
        "type": "object"
      },
...
      "StageKey": {
        "description": "Collection and analysis stages, in execution order.",
        "enum": [
          "validate",
          "dns",
          "robots",
          "http_probe",
          "tls",
          "browser_launch",
          "navigate",
          "dom_capture",
          "runtime_capture",
          "style_capture",
          "perf_capture",
          "network_capture",
          "a11y_capture",
          "responsive_probe",
          "screenshot",
          "analyze",
          "assemble"
        ],
        "title": "StageKey",
        "type": "string"
      },
      "StageProgress": {
        "additionalProperties": false,
        "description": "Progress derived from stages that actually completed.\n\nThere is no interpolation and no time-based estimate here on purpose: the requirement\nis honest progress, and the simplest way to guarantee it is to have no mechanism\ncapable of faking it.",
        "properties": {
          "completed_weight": {
            "default": 0,
            "title": "Completed Weight",
            "type": "integer"
          },
          "current_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageKey"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_stage_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage Label"
          },
          "stages_completed": {
            "default": 0,
            "title": "Stages Completed",
            "type": "integer"
          },
          "stages_total": {
            "default": 0,
            "title": "Stages Total",
            "type": "integer"
          },
          "total_weight": {
            "default": 0,
            "title": "Total Weight",
            "type": "integer"
          }
        },
        "title": "StageProgress",
        "type": "object"
      },
      "StageRun": {
        "additionalProperties": false,
        "properties": {
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Detail"
          },
          "key": {
            "$ref": "#/components/schemas/StageKey"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "skip_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skip Reason"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/StageStatus"
          }
        },
        "required": [
          "key",
          "label",
          "status"
        ],
        "title": "StageRun",
        "type": "object"
      },
      "StageStatus": {
        "enum": [
          "pending",
          "running",
          "completed",
          "failed",
          "skipped"
        ],
        "title": "StageStatus",
        "type": "string"
      },
...
      "TargetInfo": {
        "additionalProperties": false,
        "description": "What we ended up analyzing, which is not always what was requested.",
        "properties": {
          "document_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Title"
          },
          "final_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Url"
          },
          "host": {
            "title": "Host",
            "type": "string"
          },
          "http_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Http Status"
          },
          "normalized_url": {
            "title": "Normalized Url",
            "type": "string"
          },
          "port": {
            "title": "Port",
            "type": "integer"
          },
          "redirect_chain": {
            "items": {
              "$ref": "#/components/schemas/RedirectHop"
            },
            "title": "Redirect Chain",
            "type": "array"
          },
          "requested_url": {
            "title": "Requested Url",
            "type": "string"
          },
          "resolved_ips": {
            "items": {
              "type": "string"
            },
            "title": "Resolved Ips",
            "type": "array"
          },
          "robots": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RobotsObservation"
              },
              {
                "type": "null"
              }
            ]
          },
          "scheme": {
            "title": "Scheme",
            "type": "string"
          }
        },
        "required": [
          "requested_url",
          "normalized_url",
          "host",
          "port",
          "scheme"
        ],
        "title": "TargetInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Viewport": {
        "additionalProperties": false,
        "properties": {
          "height": {
            "default": 900,
            "maximum": 2160.0,
            "minimum": 320.0,
            "title": "Height",
            "type": "integer"
          },
          "width": {
            "default": 1440,
            "maximum": 3840.0,
            "minimum": 320.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "title": "Viewport",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Evidence-based website technical intelligence.\n\nDetection is deterministic and never performed by an AI model. Every asserted fact carries the\nevidence that supports it, and anything that cannot be established from observation is reported\nas not detected, not determinable, or unable to verify rather than guessed.\n\nAnalysis is passive: WebLens observes what a normal visit reveals. It does not test\nauthentication, submit forms, fuzz inputs, or attempt to bypass access controls.",
    "title": "WebLens API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scans": {
      "post": {
        "operationId": "create_scan_api_v1_scans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit a URL for analysis",
        "tags": [
          "scans"
        ]
      }
    },
    "/api/v1/scans/{scan_id}": {
      "delete": {
        "description": "Called by the client once the result is stored in IndexedDB.\n\nIdempotent: an unknown id also returns 204, because the desired end state - no server copy -\nholds either way.",
        "operationId": "delete_scan_api_v1_scans__scan_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release the server-side copy of a scan",
        "tags": [
          "scans"
        ]
      },
      "get": {
        "operationId": "get_scan_api_v1_scans__scan_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanJobState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scan status and stage progress",
        "tags": [
          "scans"
        ]
      }
    },
    "/api/v1/scans/{scan_id}/events": {
      "get": {
        "operationId": "scan_events_api_v1_scans__scan_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Server-sent events for scan progress",
        "tags": [
          "scans"
        ]
      }
    },
    "/api/v1/scans/{scan_id}/result": {
      "get": {
        "operationId": "get_result_api_v1_scans__scan_id__result_get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The scan has not finished yet."
          },
          "410": {
            "description": "The result was released after the retention window."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Structured analysis result",
        "tags": [
          "scans"
        ]
      }
    },
...
  }
}